requirements.txt
pyspark==3.5.0
psycopg2-binary==2.9.9 